+++ src/radiation/source_terms_multi_group.hpp
