+++ src/radiation/source_terms_single_group.hpp
